+++ src/main/java/lotto/util/LottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.util;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
